requirements.txt
mysql-connector-python
yara-python
smalisca
